production_app/ensembl/datacheck/restplus.py
import logging
import traceback

from flask_restplus import Api
from production_app import settings
from sqlalchemy.orm.exc import NoResultFound

log = logging.getLogger(__name__)

api = Api(version='1.0', title='Ensembl Production Datacheck App',
          description='datacheck app to perform datachecks on given DB')


@api.errorhandler
def default_error_handler(e):
    message = 'An unhandled exception occurred.'
    log.exception(message)

    if not settings.FLASK_DEBUG:
        return {'message': message}, 500


@api.errorhandler(NoResultFound)
def database_not_found_error_handler(e):
    log.warning(traceback.format_exc())
    return {'message': 'could not found requested url.'}, 404
