requirements.txt
#git+git://github.com/Ensembl/ensembl-prodinf-core@master#egg=ensembl_prodinf
amqp==2.5.0
celery==4.3.0
#certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
docutils==0.14
Flask==1.1.1
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
idna==2.8
py==1.8.0
pyasn1==0.4.5
PyMySQL==0.9.3
pyparsing==2.4.1.1
pyrsistent==0.15.3
python-dateutil==2.8.0
